Fixed an issue where transform labels were not being assigned correctly in distance outputs.
The problem occurred when the order of Transforms in the Hash Dataframe didn't match the order presented by the LabelEncoder (which does a sort when fitting labels).

This is now handled by ordering columns based on how they appear in the hash Dataframe. Still seems to required Intra-distances to be done bfore inter-, though.