Updated readme, fixed a typo
README now has a caveat section. In testing with a student we ran in to a weird issue where original images from the phone were not matching well with images edited on the phone. It turned out to be a rotation issue, and since the OS often corrects for rotation in the preview it may not be obvious what is happening.